+++ src/DotRecast.Detour.Extras/Unity/Astar/ZipBinaryReader.cs
@@ -31,7 +31,7 @@ namespace DotRecast.Detour.Extras.Unity.Astar
             ZipArchiveEntry graphReferences = file.GetEntry(filename);
             using var entryStream = graphReferences.Open();
             using var br = new BinaryReader(entryStream);
-            RcByteBuffer buffer = IOUtils.ToByteBuffer(br);
+            RcByteBuffer buffer = RcIO.ToByteBuffer(br);
             buffer.Order(RcByteOrder.LITTLE_ENDIAN);
             return buffer;
         }