+++ src/DotRecast.Detour.TileCache/Io/DtTileCacheLayerHeaderWriter.cs
@@ -24,34 +24,34 @@ using DotRecast.Detour.Io;
 
 namespace DotRecast.Detour.TileCache.Io
 {
-    public class DtTileCacheLayerHeaderWriter : DtWriter
+    public class DtTileCacheLayerHeaderWriter
     {
         public void Write(BinaryWriter stream, DtTileCacheLayerHeader header, RcByteOrder order, bool cCompatibility)
         {
-            Write(stream, header.magic, order);
-            Write(stream, header.version, order);
-            Write(stream, header.tx, order);
-            Write(stream, header.ty, order);
-            Write(stream, header.tlayer, order);
-            
-            Write(stream, header.bmin.X, order);
-            Write(stream, header.bmin.Y, order);
-            Write(stream, header.bmin.Z, order);
-            Write(stream, header.bmax.X, order);
-            Write(stream, header.bmax.Y, order);
-            Write(stream, header.bmax.Z, order);
+            RcIO.Write(stream, header.magic, order);
+            RcIO.Write(stream, header.version, order);
+            RcIO.Write(stream, header.tx, order);
+            RcIO.Write(stream, header.ty, order);
+            RcIO.Write(stream, header.tlayer, order);
 
-            Write(stream, (short)header.hmin, order);
-            Write(stream, (short)header.hmax, order);
-            Write(stream, (byte)header.width);
-            Write(stream, (byte)header.height);
-            Write(stream, (byte)header.minx);
-            Write(stream, (byte)header.maxx);
-            Write(stream, (byte)header.miny);
-            Write(stream, (byte)header.maxy);
+            RcIO.Write(stream, header.bmin.X, order);
+            RcIO.Write(stream, header.bmin.Y, order);
+            RcIO.Write(stream, header.bmin.Z, order);
+            RcIO.Write(stream, header.bmax.X, order);
+            RcIO.Write(stream, header.bmax.Y, order);
+            RcIO.Write(stream, header.bmax.Z, order);
+
+            RcIO.Write(stream, (short)header.hmin, order);
+            RcIO.Write(stream, (short)header.hmax, order);
+            RcIO.Write(stream, (byte)header.width);
+            RcIO.Write(stream, (byte)header.height);
+            RcIO.Write(stream, (byte)header.minx);
+            RcIO.Write(stream, (byte)header.maxx);
+            RcIO.Write(stream, (byte)header.miny);
+            RcIO.Write(stream, (byte)header.maxy);
             if (cCompatibility)
             {
-                Write(stream, (short)0, order); // C struct padding
+                RcIO.Write(stream, (short)0, order); // C struct padding
             }
         }
     }